Clean SamplesUtils
