+++ docs/samples/Microsoft.ML.Samples/Dynamic/NgramExtraction.cs
@@ -14,7 +14,7 @@ namespace Samples.Dynamic
             var ml = new MLContext();
 
             // Get a small dataset as an IEnumerable and convert to IDataView.
-            IEnumerable<Microsoft.ML.SamplesUtils.DatasetUtils.SampleSentimentData> data = Microsoft.ML.SamplesUtils.DatasetUtils.GetSentimentData();
+            IEnumerable<SampleSentimentData> data = GetSentimentData();
             var trainData = ml.Data.LoadFromEnumerable(data);
 
             // Preview of the data.
@@ -70,6 +70,27 @@ namespace Samples.Dynamic
             // 'B' - 1 'B|e' - 1 'e' - 6 'e|s' - 1 's' - 1 's|t' - 1 't' - 1 't|<?>' - 1 '<?>' - 4 '<?>|g' - 1 ...
             // 'e' - 1 '<?>' - 2 'd' - 1 '=' - 4 '=|=' - 2 '=|R' - 1 'R' - 1 'R|U' - 1 'U' - 1 'U|D' - 1 'D' - 2 ...
             // 'B' - 0 'B|e' - 0 'e' - 6 'e|s' - 1 's' - 3 's|t' - 1 't' - 6 't|<?>' - 2 '<?>' - 9 '<?>|g' - 2 ...
+        }
+
+        /// <summary>
+        /// A dataset that contains a tweet and the sentiment assigned to that tweet: 0 - negative and 1 - positive sentiment.
+        /// </summary>
+        public class SampleSentimentData
+        {
+            public bool Sentiment { get; set; }
+            public string SentimentText { get; set; }
+        }
+
+        /// <summary>
+        /// Returns a sample of the sentiment dataset.
+        /// </summary>
+        public static IEnumerable<SampleSentimentData> GetSentimentData()
+        {
+            var data = new List<SampleSentimentData>();
+            data.Add(new SampleSentimentData { Sentiment = true, SentimentText = "Best game I've ever played." });
+            data.Add(new SampleSentimentData { Sentiment = false, SentimentText = "==RUDE== Dude, 2" });
+            data.Add(new SampleSentimentData { Sentiment = true, SentimentText = "Until the next game, this is the best Xbox game!" });
+            return data;
         }
     }
 }
