+++ src/Microsoft.ML.SamplesUtils/SamplesDatasetUtils.cs
@@ -55,69 +55,10 @@ namespace Microsoft.ML.SamplesUtils
         }
 
         /// <summary>
-        /// A class to hold the raw housing regression rows.
+        /// Downloads the adult dataset from the ML.NET repo.
         /// </summary>
-        public sealed class HousingRegression
-        {
-            public float MedianHomeValue { get; set; }
-            public float CrimesPerCapita { get; set; }
-            public float PercentResidental { get; set; }
-            public float PercentNonRetail { get; set; }
-            public float CharlesRiver { get; set; }
-            public float NitricOxides { get; set; }
-            public float RoomsPerDwelling { get; set; }
-            public float PercentPre40s { get; set; }
-            public float EmploymentDistance { get; set; }
-            public float HighwayDistance { get; set; }
-            public float TaxRate { get; set; }
-            public float TeacherRatio { get; set; }
-        }
-
-        /// <summary>
-        /// Downloads the wikipedia detox dataset from the ML.NET repo.
-        /// </summary>
-        public static string[] DownloadSentimentDataset()
-        {
-            var trainFile = Download("https://raw.githubusercontent.com/dotnet/machinelearning/76cb2cdf5cc8b6c88ca44b8969153836e589df04/test/data/wikipedia-detox-250-line-data.tsv", "sentiment.tsv");
-            var testFile = Download("https://raw.githubusercontent.com/dotnet/machinelearning/76cb2cdf5cc8b6c88ca44b8969153836e589df04/test/data/wikipedia-detox-250-line-test.tsv", "sentimenttest.tsv");
-            return new[] { trainFile, testFile };
-        }
-
-            /// <summary>
-            /// Downloads the adult dataset from the ML.NET repo.
-            /// </summary>
-            public static string DownloadAdultDataset()
-            => Download("https://raw.githubusercontent.com/dotnet/machinelearning/244a8c2ac832657af282aa312d568211698790aa/test/data/adult.train", "adult.txt");
-
-        /// <summary>
-        /// Downloads the  wikipedia detox dataset and featurizes it to be suitable for sentiment classification tasks.
-        /// </summary>
-        /// <param name="mlContext"><see cref="MLContext"/> used for data loading and processing.</param>
-        /// <returns>Featurized train and test dataset.</returns>
-        public static IDataView[] LoadFeaturizedSentimentDataset(MLContext mlContext)
-        {
-            // Download the files
-            var dataFiles = DownloadSentimentDataset();
-
-            // Define the columns to load
-            var loader = mlContext.Data.CreateTextLoader(
-                columns: new[]
-                    {
-                        new TextLoader.Column("Sentiment", DataKind.Boolean, 0),
-                        new TextLoader.Column("SentimentText", DataKind.String, 1)
-                    },
-                hasHeader: true
-            );
-
-            // Create data featurizing pipeline
-            var pipeline = mlContext.Transforms.Text.FeaturizeText("Features", "SentimentText");
-
-            var data = loader.Load(dataFiles[0]);
-            var model = pipeline.Fit(data);
-            var featurizedDataTrain = model.Transform(data);
-            var featurizedDataTest = model.Transform(loader.Load(dataFiles[1]));
-            return new[] { featurizedDataTrain, featurizedDataTest };
-        }
+        public static string DownloadAdultDataset()
+        => Download("https://raw.githubusercontent.com/dotnet/machinelearning/244a8c2ac832657af282aa312d568211698790aa/test/data/adult.train", "adult.txt");
 
         /// <summary>
         /// Downloads the Adult UCI dataset and featurizes it to be suitable for classification tasks.
@@ -176,43 +117,6 @@ namespace Microsoft.ML.SamplesUtils
             var data = loader.Load(dataFile);
             var featurizedData = pipeline.Fit(data).Transform(data);
             return featurizedData;
-        }
-
-        public static string DownloadMslrWeb10k()
-        {
-            var fileName = "MSLRWeb10KTrain10kRows.tsv";
-            if (!File.Exists(fileName))
-                Download("https://tlcresources.blob.core.windows.net/datasets/MSLR-WEB10K/MSLR-WEB10K%2BFold1.TRAIN.SMALL_10k-rows.tsv", fileName);
-            return fileName;
-        }
-
-        public static IDataView LoadFeaturizedMslrWeb10kDataset(MLContext mlContext)
-        {
-            // Download the training and validation files.
-            string dataFile = DownloadMslrWeb10k();
-
-            // Create the loader to load the data.
-            var loader = mlContext.Data.CreateTextLoader(
-                columns: new[]
-                {
-                    new TextLoader.Column("Label", DataKind.Single, 0),
-                    new TextLoader.Column("GroupId", DataKind.String, 1),
-                    new TextLoader.Column("Features", DataKind.Single, new[] { new TextLoader.Range(2, 138) })
-                }
-            );
-
-            // Load the raw dataset.
-            var data = loader.Load(dataFile);
-
-            // Create the featurization pipeline. First, hash the GroupId column.
-            var pipeline = mlContext.Transforms.Conversion.Hash("GroupId")
-                // Replace missing values in Features column with the default replacement value for its type.
-                .Append(mlContext.Transforms.ReplaceMissingValues("Features"));
-
-            // Fit the pipeline and transform the dataset.
-            var transformedData = pipeline.Fit(data).Transform(data);
-
-            return transformedData;
         }
 
         /// <summary>
@@ -281,231 +185,6 @@ namespace Microsoft.ML.SamplesUtils
             }
 
             return dataFile;
-        }
-
-        /// <summary>
-        /// A simple set of features that help generate the Target column, according to a function.
-        /// Used for the transformers/estimators working on numeric data.
-        /// </summary>
-        public class SampleInput
-        {
-            public float Feature0 { get; set; }
-            public float Feature1 { get; set; }
-            public float Feature2 { get; set; }
-            public float Feature3 { get; set; }
-            public float Target { get; set; }
-        }
-
-        /// <summary>
-        /// Returns a sample of a numeric dataset.
-        /// </summary>
-        public static IEnumerable<SampleInput> GetInputData()
-        {
-            var data = new List<SampleInput>();
-            data.Add(new SampleInput { Feature0 = -2.75f, Feature1 = 0.77f, Feature2 = -0.61f, Feature3 = 0.14f, Target = 140.66f });
-            data.Add(new SampleInput { Feature0 = -0.61f, Feature1 = -0.37f, Feature2 = -0.12f, Feature3 = 0.55f, Target = 148.12f });
-            data.Add(new SampleInput { Feature0 = -0.85f, Feature1 = -0.91f, Feature2 = 1.81f, Feature3 = 0.02f, Target = 402.20f });
-
-            return data;
-        }
-
-        /// <summary>
-        /// A dataset that contains a tweet and the sentiment assigned to that tweet: 0 - negative and 1 - positive sentiment.
-        /// </summary>
-        public class SampleSentimentData
-        {
-            public bool Sentiment { get; set; }
-            public string SentimentText { get; set; }
-        }
-
-        /// <summary>
-        /// Returns a sample of the sentiment dataset.
-        /// </summary>
-        public static IEnumerable<SampleSentimentData> GetSentimentData()
-        {
-            var data = new List<SampleSentimentData>();
-            data.Add(new SampleSentimentData { Sentiment = true, SentimentText = "Best game I've ever played." });
-            data.Add(new SampleSentimentData { Sentiment = false, SentimentText = "==RUDE== Dude, 2" });
-            data.Add(new SampleSentimentData { Sentiment = true, SentimentText = "Until the next game, this is the best Xbox game!" });
-
-            return data;
-        }
-
-        /// <summary>
-        /// A dataset that contains one column with two set of keys assigned to a body of text: Review and ReviewReverse.
-        /// The dataset will be used to classify how accurately the keys are assigned to the text.
-        /// </summary>
-        public class SampleTopicsData
-        {
-            public string Review { get; set; }
-            public string ReviewReverse { get; set; }
-            public bool Label { get; set; }
-        }
-
-        /// <summary>
-        /// Returns a sample of the topics dataset.
-        /// </summary>
-        public static IEnumerable<SampleTopicsData> GetTopicsData()
-        {
-            var data = new List<SampleTopicsData>();
-            data.Add(new SampleTopicsData { Review = "animals birds cats dogs fish horse", ReviewReverse = "radiation galaxy universe duck", Label = true });
-            data.Add(new SampleTopicsData { Review = "horse birds house fish duck cats", ReviewReverse = "space galaxy universe radiation", Label = false });
-            data.Add(new SampleTopicsData { Review = "car truck driver bus pickup", ReviewReverse = "bus pickup", Label = true });
-            data.Add(new SampleTopicsData { Review = "car truck driver bus pickup horse", ReviewReverse = "car truck", Label = false });
-
-            return data;
-        }
-
-        public class SampleTemperatureData
-        {
-            public DateTime Date { get; set; }
-            public float Temperature { get; set; }
-        }
-
-        public class SampleTemperatureDataWithLatitude
-        {
-            public float Latitude { get; set; }
-            public DateTime Date { get; set; }
-            public float Temperature { get; set; }
-        }
-
-        /// <summary>
-        /// Get a fake temperature dataset.
-        /// </summary>
-        /// <param name="exampleCount">The number of examples to return.</param>
-        /// <returns>An enumerable of <see cref="SampleTemperatureData"/>.</returns>
-        public static IEnumerable<SampleTemperatureData> GetSampleTemperatureData(int exampleCount)
-        {
-            var rng = new Random(1234321);
-            var date = new DateTime(2012, 1, 1);
-            float temperature = 39.0f;
-
-            for (int i = 0; i < exampleCount; i++)
-            {
-                date = date.AddDays(1);
-                temperature += rng.Next(-5, 5);
-                yield return new SampleTemperatureData { Date = date, Temperature = temperature };
-            }
-        }
-
-        /// <summary>
-        /// Represents the column of the infertility dataset.
-        /// </summary>
-        public class SampleInfertData
-        {
-            public int RowNum { get; set; }
-            public string Education { get; set; }
-            public float Age { get; set; }
-            public float Parity { get; set; }
-            public float Induced { get; set; }
-            public float Case { get; set; }
-
-            public float Spontaneous { get; set; }
-            public float Stratum { get; set; }
-            public float PooledStratum { get; set; }
-        }
-
-        /// <summary>
-        /// Returns a few rows of the infertility dataset.
-        /// </summary>
-        public static IEnumerable<SampleInfertData> GetInfertData()
-        {
-            var data = new List<SampleInfertData>();
-            data.Add(new SampleInfertData
-            {
-                RowNum = 0,
-                Education = "0-5yrs",
-                Age = 26,
-                Parity = 6,
-                Induced = 1,
-                Case = 1,
-                Spontaneous = 2,
-                Stratum = 1,
-                PooledStratum = 3
-            });
-            data.Add(new SampleInfertData
-            {
-                RowNum = 1,
-                Education = "0-5yrs",
-                Age = 42,
-                Parity = 1,
-                Induced = 1,
-                Case = 1,
-                Spontaneous = 0,
-                Stratum = 2,
-                PooledStratum = 1
-            });
-            data.Add(new SampleInfertData
-            {
-                RowNum = 2,
-                Education = "12+yrs",
-                Age = 39,
-                Parity = 6,
-                Induced = 2,
-                Case = 1,
-                Spontaneous = 0,
-                Stratum = 3,
-                PooledStratum = 4
-            });
-            data.Add(new SampleInfertData
-            {
-                RowNum = 3,
-                Education = "0-5yrs",
-                Age = 34,
-                Parity = 4,
-                Induced = 2,
-                Case = 1,
-                Spontaneous = 0,
-                Stratum = 4,
-                PooledStratum = 2
-            });
-            data.Add(new SampleInfertData
-            {
-                RowNum = 4,
-                Education = "6-11yrs",
-                Age = 35,
-                Parity = 3,
-                Induced = 1,
-                Case = 1,
-                Spontaneous = 1,
-                Stratum = 5,
-                PooledStratum = 32
-            });
-            return data;
-        }
-
-        public class SampleVectorOfNumbersData
-        {
-            [VectorType(10)]
-
-            public float[] Features { get; set; }
-        }
-
-        /// <summary>
-        /// Returns a few rows of the infertility dataset.
-        /// </summary>
-        public static IEnumerable<SampleVectorOfNumbersData> GetVectorOfNumbersData()
-        {
-            var data = new List<SampleVectorOfNumbersData>();
-            data.Add(new SampleVectorOfNumbersData { Features = new float[10] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 } });
-            data.Add(new SampleVectorOfNumbersData { Features = new float[10] { 1, 2, 3, 4, 5, 6, 7, 8, 9, 0 } });
-            data.Add(new SampleVectorOfNumbersData
-            {
-                Features = new float[10] { 2, 3, 4, 5, 6, 7, 8, 9, 0, 1 }
-            });
-            data.Add(new SampleVectorOfNumbersData
-            {
-                Features = new float[10] { 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, }
-            });
-            data.Add(new SampleVectorOfNumbersData
-            {
-                Features = new float[10] { 5, 6, 7, 8, 9, 0, 1, 2, 3, 4 }
-            });
-            data.Add(new SampleVectorOfNumbersData
-            {
-                Features = new float[10] { 6, 7, 8, 9, 0, 1, 2, 3, 4, 5 }
-            });
-            return data;
         }
 
         private const int _simpleBinaryClassSampleFeatureLength = 10;
@@ -607,61 +286,6 @@ namespace Microsoft.ML.SamplesUtils
             return data;
         }
 
-        public class FfmExample
-        {
-            public bool Label;
-
-            [VectorType(_simpleBinaryClassSampleFeatureLength)]
-            public float[] Field0;
-
-            [VectorType(_simpleBinaryClassSampleFeatureLength)]
-            public float[] Field1;
-
-            [VectorType(_simpleBinaryClassSampleFeatureLength)]
-            public float[] Field2;
-        }
-
-        public static IEnumerable<FfmExample> GenerateFfmSamples(int exampleCount)
-        {
-            var rnd = new Random(0);
-            var data = new List<FfmExample>();
-            for (int i = 0; i < exampleCount; ++i)
-            {
-                // Initialize an example with a random label and an empty feature vector.
-                var sample = new FfmExample()
-                {
-                    Label = rnd.Next() % 2 == 0,
-                    Field0 = new float[_simpleBinaryClassSampleFeatureLength],
-                    Field1 = new float[_simpleBinaryClassSampleFeatureLength],
-                    Field2 = new float[_simpleBinaryClassSampleFeatureLength]
-                };
-                // Fill feature vector according the assigned label.
-                for (int j = 0; j < 10; ++j)
-                {
-                    var value0 = (float)rnd.NextDouble();
-                    // Positive class gets larger feature value.
-                    if (sample.Label)
-                        value0 += 0.2f;
-                    sample.Field0[j] = value0;
-
-                    var value1 = (float)rnd.NextDouble();
-                    // Positive class gets smaller feature value.
-                    if (sample.Label)
-                        value1 -= 0.2f;
-                    sample.Field1[j] = value1;
-
-                    var value2 = (float)rnd.NextDouble();
-                    // Positive class gets larger feature value.
-                    if (sample.Label)
-                        value2 += 0.8f;
-                    sample.Field2[j] = value2;
-                }
-
-                data.Add(sample);
-            }
-            return data;
-        }
-
         /// <summary>
         /// feature vector's length in <see cref="MulticlassClassificationExample"/>.
         /// </summary>
@@ -723,50 +347,6 @@ namespace Microsoft.ML.SamplesUtils
                 examples.Add(example);
             }
             return examples;
-        }
-
-        // The following variables defines the shape of a matrix. Its shape is _synthesizedMatrixRowCount-by-_synthesizedMatrixColumnCount.
-        // Because in ML.NET key type's minimal value is zero, the first row index is always zero in C# data structure (e.g., MatrixColumnIndex=0
-        // and MatrixRowIndex=0 in MatrixElement below specifies the value at the upper-left corner in the training matrix). If user's row index
-        // starts with 1, their row index 1 would be mapped to the 2nd row in matrix factorization module and their first row may contain no values.
-        // This behavior is also true to column index.
-        private const uint _synthesizedMatrixFirstColumnIndex = 1;
-        private const uint _synthesizedMatrixFirstRowIndex = 1;
-        private const uint _synthesizedMatrixColumnCount = 60;
-        private const uint _synthesizedMatrixRowCount = 100;
-
-        // A data structure used to encode a single value in matrix
-        public class MatrixElement
-        {
-            // Matrix column index is at most _synthesizedMatrixColumnCount + _synthesizedMatrixFirstColumnIndex.
-            [KeyType(_synthesizedMatrixColumnCount + _synthesizedMatrixFirstColumnIndex)]
-            public uint MatrixColumnIndex;
-            // Matrix row index is at most _synthesizedMatrixRowCount + _synthesizedMatrixFirstRowIndex.
-            [KeyType(_synthesizedMatrixRowCount + _synthesizedMatrixFirstRowIndex)]
-            public uint MatrixRowIndex;
-            // The value at the column MatrixColumnIndex and row MatrixRowIndex.
-            public float Value;
-        }
-
-        // A data structure used to encode prediction result. Comparing with MatrixElement, The field Value in MatrixElement is
-        // renamed to Score because Score is the default name of matrix factorization's output.
-        public class MatrixElementForScore
-        {
-            [KeyType(_synthesizedMatrixColumnCount + _synthesizedMatrixFirstColumnIndex)]
-            public uint MatrixColumnIndex;
-            [KeyType(_synthesizedMatrixRowCount + _synthesizedMatrixFirstRowIndex)]
-            public uint MatrixRowIndex;
-            public float Score;
-        }
-
-        // Create an in-memory matrix as a list of tuples (column index, row index, value).
-        public static List<MatrixElement> GetRecommendationData()
-        {
-            var dataMatrix = new List<MatrixElement>();
-            for (uint i = _synthesizedMatrixFirstColumnIndex; i < _synthesizedMatrixFirstColumnIndex + _synthesizedMatrixColumnCount; ++i)
-                for (uint j = _synthesizedMatrixFirstRowIndex; j < _synthesizedMatrixFirstRowIndex + _synthesizedMatrixRowCount; ++j)
-                    dataMatrix.Add(new MatrixElement() { MatrixColumnIndex = i, MatrixRowIndex = j, Value = (i + j) % 5 });
-            return dataMatrix;
         }
     }
 }
