+++ test/Microsoft.ML.Tests/TrainerEstimators/FAFMEstimator.cs
@@ -2,9 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
+using System.Collections.Generic;
 using Microsoft.ML.Data;
 using Microsoft.ML.RunTests;
-using Microsoft.ML.SamplesUtils;
 using Microsoft.ML.Trainers;
 using Xunit;
 
@@ -16,10 +17,10 @@ namespace Microsoft.ML.Tests.TrainerEstimators
         public void FfmBinaryClassificationWithoutArguments()
         {
             var mlContext = new MLContext(seed: 0);
-            var data = DatasetUtils.GenerateFfmSamples(500);
+            var data = GenerateFfmSamples(500);
             var dataView = mlContext.Data.LoadFromEnumerable(data);
 
-            var pipeline = mlContext.Transforms.CopyColumns(DefaultColumnNames.Features, nameof(DatasetUtils.FfmExample.Field0))
+            var pipeline = mlContext.Transforms.CopyColumns(DefaultColumnNames.Features, nameof(FfmExample.Field0))
                 .Append(mlContext.BinaryClassification.Trainers.FieldAwareFactorizationMachine());
 
             var model = pipeline.Fit(dataView);
@@ -37,14 +38,14 @@ namespace Microsoft.ML.Tests.TrainerEstimators
         public void FfmBinaryClassificationWithAdvancedArguments()
         {
             var mlContext = new MLContext(seed: 0);
-            var data = DatasetUtils.GenerateFfmSamples(500);
+            var data = GenerateFfmSamples(500);
             var dataView = mlContext.Data.LoadFromEnumerable(data);
 
             var ffmArgs = new FieldAwareFactorizationMachineTrainer.Options();
 
             // Customized the field names.
-            ffmArgs.FeatureColumnName = nameof(DatasetUtils.FfmExample.Field0); // First field.
-            ffmArgs.ExtraFeatureColumns = new[]{ nameof(DatasetUtils.FfmExample.Field1), nameof(DatasetUtils.FfmExample.Field2) };
+            ffmArgs.FeatureColumnName = nameof(FfmExample.Field0); // First field.
+            ffmArgs.ExtraFeatureColumns = new[] { nameof(FfmExample.Field1), nameof(FfmExample.Field2) };
 
             var pipeline = mlContext.BinaryClassification.Trainers.FieldAwareFactorizationMachine(ffmArgs);
 
@@ -65,9 +66,10 @@ namespace Microsoft.ML.Tests.TrainerEstimators
             var data = new TextLoader(Env, GetFafmBCLoaderArgs())
                     .Load(GetDataPath(TestDatasets.breastCancer.trainFilename));
 
-            var ffmArgs = new FieldAwareFactorizationMachineTrainer.Options {
+            var ffmArgs = new FieldAwareFactorizationMachineTrainer.Options
+            {
                 FeatureColumnName = "Feature1", // Features from the 1st field.
-                ExtraFeatureColumns = new[] { "Feature2", "Feature3",  "Feature4" }, // 2nd field's feature column, 3rd field's feature column, 4th field's feature column.
+                ExtraFeatureColumns = new[] { "Feature2", "Feature3", "Feature4" }, // 2nd field's feature column, 3rd field's feature column, 4th field's feature column.
                 Shuffle = false,
                 NumberOfIterations = 3,
                 LatentDimension = 7,
@@ -97,6 +99,63 @@ namespace Microsoft.ML.Tests.TrainerEstimators
                     new TextLoader.Column("Label", DataKind.Boolean, 0)
                 }
             };
+        }
+
+        private const int _simpleBinaryClassSampleFeatureLength = 10;
+
+        private class FfmExample
+        {
+            public bool Label;
+
+            [VectorType(_simpleBinaryClassSampleFeatureLength)]
+            public float[] Field0;
+
+            [VectorType(_simpleBinaryClassSampleFeatureLength)]
+            public float[] Field1;
+
+            [VectorType(_simpleBinaryClassSampleFeatureLength)]
+            public float[] Field2;
+        }
+
+        private static IEnumerable<FfmExample> GenerateFfmSamples(int exampleCount)
+        {
+            var rnd = new Random(0);
+            var data = new List<FfmExample>();
+            for (int i = 0; i < exampleCount; ++i)
+            {
+                // Initialize an example with a random label and an empty feature vector.
+                var sample = new FfmExample()
+                {
+                    Label = rnd.Next() % 2 == 0,
+                    Field0 = new float[_simpleBinaryClassSampleFeatureLength],
+                    Field1 = new float[_simpleBinaryClassSampleFeatureLength],
+                    Field2 = new float[_simpleBinaryClassSampleFeatureLength]
+                };
+                // Fill feature vector according the assigned label.
+                for (int j = 0; j < 10; ++j)
+                {
+                    var value0 = (float)rnd.NextDouble();
+                    // Positive class gets larger feature value.
+                    if (sample.Label)
+                        value0 += 0.2f;
+                    sample.Field0[j] = value0;
+
+                    var value1 = (float)rnd.NextDouble();
+                    // Positive class gets smaller feature value.
+                    if (sample.Label)
+                        value1 -= 0.2f;
+                    sample.Field1[j] = value1;
+
+                    var value2 = (float)rnd.NextDouble();
+                    // Positive class gets larger feature value.
+                    if (sample.Label)
+                        value2 += 0.8f;
+                    sample.Field2[j] = value2;
+                }
+
+                data.Add(sample);
+            }
+            return data;
         }
     }
 }
